found the README!
ok so I thought it wasn't working because I couldn't find the newly created README even though I was getting the success message in the terminal. But what was happening was it was overwriting the README that I had already created in the repo! Gosh do I feel dumb! Dumb and very very relieved.